Integrated support for per-interval metric reporting. It requires cooperation from the generator subclasses and the application-specific Operation base classes e.g. the OlioOperation base-class for more specific OlioOperations. Also load intervals (LoadProfile instances must be given a name).
Concrete generators must save the current load profile in their _latestLoadProfile
during nextRequest and application-specific operation base classes must call
Generator::getLatestLoadProfile() and save the result (if non-null) in their
_generatedDuring member variable during Operation::prepare(Generator).
When an operation executes it creates an OperationExecution instance that
contains the value stored in Operation::_generatedDuring. Once the
OperationExecution makes its way to the Scoreboard, the Scoreboard
will look at the OperationExecution::_generatedDuring value and use
the name on the loadprofile as the interval name. If the name on
the loadprofile is null or empty-string, it will be ignored, i.e.,
no per-interval results will be collected.