Updated the sampling strategy interface to include methods for computing the sample mean, sample standard distribution and a t-value (given the observed population mean). We can use these quantities to determine whether the sampling interval is adequate (i.e., it results in the collection of sufficient samples to determine how tight the estimates of higher percentiles of the quantities measured, e.g., response times and wait times, are.
Updated PoissonSamplingStrategy to compute these extra stats
(sample mean, sample standard deviation, t-value).

Updated the WaitTimeSummary and OperationSummary to ask the
ISamplingStrategy implementations for these extra stats.

Updated the Scoreboard to display these extra stats next to
the operation and wait time final results.

Updated burstUrl.py to allow passing the sampling interval as a parameter
--sampleinterval <value> from the command line.

Updated the BurstUrlProfileCreator to look for the sampling interval
as a configuration parameter.