Added pre- and post- execute hooks to the operation to allow for activities before and after operation execution timing starts. In the context of Scadr we're using these hooks to keep track of the number of requests outstanding per-server.
Updated ScadrGenerator and ScadrScenarioTrack classes such that we can
find the least loaded server based on the number of outstanding requests
(request-gating). This way, load generator threads will always go to the
least loaded machine instead of hammering (and getting backed up behind)
an already overloaded machine.

Added a ScadrNullOperation to test the contention for the lock associated
with tracking the number of outstanding requests per-server.

Updated the ScadrGenerator to send requests to the least loaded server
based on the traffic stats.