Updated MongoTransport to be a singleton shared across threads. Since MongoTransport wraps a Mongo client (each of which manages a connection pool) having a single MongoTransport is fine -- reducing the number of pooled connections and the number of open connections to mongodb and it's threadsafe to use it as a singleton since the underlying Mongo client that MongoTransport delegates to is threadsafe.
Added WriteConcern.Safe to Mongo client calls to update and insert. This
makes sure that GetLastError is called on a connection before it is
returned to the connection pool - this prevents any IllegalStateExceptions
that would occur due to another write happening before getLastError is called
for an earlier write on that connection.

Reduced the default number of shared threads to 10 so that in the
common case of closed loop workloads there aren't many shared threads
hanging around consuming resources.

Updated MongoGenerator and MongoUtil to use the MongoTransport singleton.