We're reverting to the default of one-mongoclient-per-thread.
Sharing a single Mongo client (even though it's threadsafe) results in
about ~6 connections per thread whereas having a Mongo client per thread
leads to about ~2 connections per thread (as seen in mongostat) so running
with larger numbers of concurrent threads (e.g. 1600+) can lead to >10000
connections in the shared Mongo client case vs. 3200+ connections in the
one-client-per-thread case. With a large number of open connections to
MongoDB there are lots of connection timeout/reset errors so the
one-client-per-thread approach is preferred for high-concurrency workloads.

For smaller numbers of users O(100) concurrent threads, throughput is ~7%
better for some object sizes with a single shared Mongo client than with
one-client-per-thread.

To work well across different concurrency levels and object sizes we're going
back to a default of one-mongoclient-per-thread.