(bug-fix:major) Wrongrly use of HTTP headers as HTTP parameters.
(change:minor) Replaced use of MultipartEntity with UrlEncodedFormEntity, since the former is only needed for file uploads.

(new:major) Use of Apache HTTP Components version 4.2.3 in place of 4.1.3. This is needed for the use of the URIBuilder in the RUBiS workload. As a side effect, Java compilation now issue a warning about the use of deprecated API in the HttpTransport class.